small update on getUser
