+++ web-app/src/components/App.js
@@ -10,7 +10,31 @@ import InternalApp from "./InternalApp";
 // React Helmets
 import { Helmet, HelmetProvider } from "react-helmet-async";
 
+// Apollo
+import GET_USER from "../graphql/queries/GET_USER";
+import { useQuery } from "@apollo/client";
+
+// Redux
+import { getUserSuccess, getUserFailure } from "../state/slices/userSlice";
+import { useDispatch } from "react-redux";
+
 function App() {
+  const dispatch = useDispatch();
+
+  const { loading } = useQuery(GET_USER, {
+    onCompleted: ({ name, email }) => {
+      // TODO: update to new method or decode from jwt_decode
+      const userInfo = { name, email };
+      dispatch(getUserSuccess(userInfo));
+    },
+    onError: (err) => {
+      console.log(err);
+      dispatch(getUserFailure());
+    },
+  });
+
+  if (loading) return <div>Initializing the App</div>;
+
   return (
     <HelmetProvider>
       <Helmet>
