+++ web-app/src/components/InternalApp.js
@@ -18,31 +18,7 @@ import Dashboard from "./pages/Dashboard";
 import PrivateRoute from "./routes/PrivateRoute";
 import PublicRoute from "./routes/PublicRoute";
 
-// Apollo
-import GET_USER from "../graphql/queries/GET_USER";
-import { useQuery } from "@apollo/client";
-
-// Redux
-import { getUserSuccess, getUserFailure } from "../state/slices/userSlice";
-import { useDispatch } from "react-redux";
-
 const Routes = () => {
-  const dispatch = useDispatch();
-
-  const { loading } = useQuery(GET_USER, {
-    onCompleted: ({ name, email }) => {
-      // TODO: update to new method or decode from jwt_decode
-      const userInfo = { name, email };
-      dispatch(getUserSuccess(userInfo));
-    },
-    onError: (err) => {
-      console.log(err);
-      dispatch(getUserFailure());
-    },
-  });
-
-  if (loading) return <div>Loading</div>;
-
   return (
     <Switch>
       <PublicRoute
